requirements.txt
requests
numpy
pyinstaller
PyQt6
ffmpeg-python
opencv-python